add custom allocator option
This is a good idea for libraries in general, and helps with tracking
memory allocated by uvwasi in Node.js.

The allocator struct format is identical to the one used by nghttp2
and related libraries, which makes things a bit easier because we’ve
added generic support for that format in the QUIC repo.