reset pointer to prevent memory corruption
The code in path_resolver.c that replaces forward slashes with
backslashes on Windows did not work. It modified a chunk of
memory immediately following the memory it was meant to update.
This was leading to crashes on Windows in the Node.js test
suite ~1-2% of the time. This commit resets the pointer to the
correct location for the calculations.

Side note: this logic was broken and the slashes weren't actually
being updated, but the tests were still passing on Windows. This
logic might not really be needed after all?