build: add uvwasi to build summary
Currently, the summary when configuring the build with cmake look like
this:

-- summary of build options:
    Install prefix:  /usr/local
    Target system:   Linux
    Compiler:
      C compiler:    /usr/lib64/ccache/cc
      CFLAGS:         -Wmaybe-uninitialized -O3

-- summary of build options:

    Install prefix:  /usr/local
    Target system:   Linux
    Compiler:
      C compiler:    /usr/lib64/ccache/cc
      CFLAGS:         -Wmaybe-uninitialized -O3
    Libuv version:   v1.38.1
    Libuv lib:       /wasm/uvwasi/build/_deps/libuv-build

-- Configuring done
-- Generating done
-- Build files have been written to: /wasm/uvwasi/build

The first summary is from libuv and the second is from this project.
This commit suggest adding uvwasi to the second to make it clear which
properties belong to which project(at first I though that the
configuration was somehow being run twice).