Fix bufused when there are more entries to be read
This commit updates the bufused value to be that of the buffer length
(buf_len) when there are more entries to be read.

Currently what is happening is that if another directory entry cannot
be written to the buffer then the current value of bufused is being
"returned", which indicates that all directory entries have been
written to the buffer which is not the case.

Fixes: https://github.com/nodejs/uvwasi/issues/174